Test with Robolectric and set up Travis.
Reorganized and cleaned up the project resources a bit, mostly
in the example project. Had to move tests out of the library
project and into the example app, as Robolectric currently has
problems with library projects, but I'll move them back when
that bug is fixed.

Updated the bundled version of stripe-java to 1.15.1.

Builds in both Eclipse and Android Studio just fine.