Fix IconTextInputLayout's reflection issues
The fields that IconTextInputLayout attempts to access
via reflection have had their names changed in the latest
Support Library upgrade, which causes an exception.

This PR fixes the exception by updating the possible
field names to include those in the latest Support Library
version.

Ideally we'll find a solution that avoids reflection, but
for the time being this works on latest Android Support.

JIRA: ANDROID-308