Create ErrorMessageTranslator and default implementation
**Summary**
ErrorMessageTranslator is an interface that a user can
implement to specify custom translations for
server-provided errors.

**Motivation**
Stripe error messages are in English. In a few places,
we show these messages on screen in dialogs. An app
developer may wish to provide their own translation of
these messages so that are legibile to their customers.

Fixes #680