Use Application context instead of Activity in Stripe.java (#783)
**Summary**
The crux of this change is making Stripe constructors use
Application context instead of Activity context.

This change also marks instance variables as `final` and
`@NonNull` where applicable.

**Motivation**
Relying on activity context for asynchronous tasks can
lead to memory leaks. We would only need an Activity
context to access the UI, which is not the case for any
of the logic in Stripe.java.

**Testing**
Interacted with all of the screens in the example app.