Add metadata field to Card (#786)
**Summary**
`metadata` is a field that exists on several Stripe objects. For this
patch, it will only be enabled for Card-based `Source` objects.

Setting `metadata` on a `Card` will ultimately populate the `metadata`
field in `SourceParams` via `SourceParams#createCardParams`, which is
passed along with the request to create a card-based `Source`.

See https://stripe.com/docs/api/metadata

This patch also does cleanup on some of the related classes (e.g.
make variables final, add nullity annotations)

**Motivation**
Being able to set `metadata` is a user-requested feature. Most of
the wiring was in place to support it, the user just needed a way
to set it on a `Card`.

**Testing**
- Added automated tests.
- In the example app, added metadata to the `Card` used in
  `PaymentMultilineActivity#saveCard`. Confirmed using the
  Dashboard's Logs feature that the metadata was included
  in the request and returned in the response.

Fixes #472