Prevent AddSourceActivity's SourceCallback from being GCed
**Summary**
`Stripe#CreateSourceTask` holds a `WeakReference` to the
`SourceCallback`. This reference was being garbage collected,
causing callback (either onSuccess or onError) to never happen.

Making the SourceCallback an instance variable seems to fix
this issue.

**Motivation**
Because the callback was never happening, the user would be
stuck on the "Add a Card" screen.

**Testing**
Tested manually on various API levels.