Add StripeError field to StripeException (#793)
**Summary**
Add a `StripeError` field to `StripeException`.
Access via `StripeException#getStripeError()`

**Motivation**
The error code is an important field for SDK users
to be able to determine why a request failed, and
allows them to better react to the failure.

For example, a `"payment_method_not_available"` code
may suggest to the developer to invite a customer to
use another payment method. Additionally, a developer
may choose to localize an user-facing error based on
this code.

Fixes #792

**Testing**
Wrote automated tests.