Improve PaymentMethodCreateParams.Card creation (#862)
When creating a `PaymentMethod` with a card, the card object
can either be created from the user's card details (e.g. card
number, CVC, expiration), or via a Stripe token (e.g. Google
Pay). `PaymentMethodCreateParams.Card.Builder`'s setters made
it confusing to the developer that there were two separate
approaches to creating a `PaymentMethodCreateParams.Card`.

With this chamge, the Builder will be used when creating with
card details, and `PaymentMethodCreateParams.Card#create` will
be used when creating with a Stripe token.

ANDROID-346