Ensure withShouldUseStripeSdk does not drop paymentMethodCreateParams
When confirming a SetupIntent or a PaymentIntent, the paymentController.startConfirmAndAuth(...) method is called which creates a ConfirmStripeIntentTask.  During creation, the given ConfirmStripeIntentParams object is transformed via its .withShouldUseStripeSdk(true) interface method.  There are two implementations of this method, one for ConfirmSetupIntentParams and one for ConfirmPaymentIntentParams.

The one for ConfirmPaymentIntentParams faithfully carries forward all of the properties, while ConfirmSetupIntentParams has an unexpected side effect of dropping the mPaymentMethodCreateParams if they are present.

The result of this makes it impossible to confirm a SetupIntent using a new payment method.  The workaround would require a user to create a payment method separately, obtain the payment method ID and then use that to confirm the SetupIntent (We shouldn't require users to do this, since, like a PaymentIntent, we should be able to specify the payment method using the PaymentMethodCreateParams during the confirm step.

If a user attempts to confirm a SetupIntent via stripe.confirmSetupIntent(..) after providing it with a new payment method (using PaymentMethodCreateParams), they will see an error like this:

```
com.stripe.android.exception.InvalidRequestException: You cannot confirm this SetupIntent because it's missing a payment method. Update the SetupIntent with a payment method and then confirm it again.; request-id: req_DRwGpA9WfR3mhw
```

Gist of logs:
https://gist.github.com/mindlapse/e4a4c2246e6e138af6e9acf646e04109