Consolidate parameter compacting logic (#1343)
The Stripe API doesn't allow for null values to be passed
as parameters, so the SDK is responsible for removing them.

The "compacting" logic (i.e. removing nulls and empty values)
was done across our codebase. This logic only needs to be
applied once, at the point that the request object is
being created.

This PR moves the logic to a single place - the `StripeRequest`
constructor.