Fix out-of-band web payment authentication
When a user enters the in-app web payment authentication flow,
is taken to another app to complete payment authentication
out-of-band (i.e. a bank app), then returns to the WebView,
make a request to the completion URL.

This resolves the issue where the bank page loaded in the
WebView doesn't detect that the user returned from the
bank app.