Convert ApiOperation from AsyncTask to Coroutines
Refactor `ApiOperation` to use coroutines to handle making
IO requests (i.e. Stripe API requests) on a background thread
(i.e. `Dispatchers.IO`) and dispatch the result to the
main thread (i.e. `MainScope()`).

`ApiOperation` takes a `workScope: CoroutineScope` param
that defaults to `Dispatchers.IO` but can be set to
`MainScope()` for testing purposes, as is shown in `StripeTest`.