Stop broadcasting error in CustomerSessionRunnableFactory
`CustomerSessionRunnableFactory` would previously report
errors both through a listener callback method and through
a `LocalBroadcastManager` broadcast.

The latter was used to show an `AlertDialog` on `StripeActivity`
subclasses. However, this is unnecessary because the alert
dialogs can be shown when the callback happens.

Also consolidate `AlertDialog` displaying logic into
`AlertDisplayer` interface and implementation.