Fix regression in CardInputWidget styling
The `android:textSize` attribute in `card_input_widget.xml` was
hardcoded in #1930, preventing styling of the `CardInputWidget`'s
EditText fields.

Introduce a base style `Stripe.Base.CardInputWidget.EditText`, and
override it with `Stripe.CardInputWidget.EditText` to allow for
user customization.

For example, in your app's `style.xml`, if you add the following:

```
<style name="Stripe.CardInputWidget.EditText" parent="Stripe.Base.CardInputWidget.EditText">
    <item name="android:textSize">22sp</item>
    <item name="android:textColor">@android:color/holo_blue_light</item>
    <item name="android:textColorHint">@android:color/holo_orange_light</item>
</style>
```

Fixes #1997