Fix backspace functionality in CardNumberEditText and ExpiryDateEditText
When backspace is pressed on `CardNumberEditText` and
`ExpiryDateEditText`, skip the `setText()` logic in
`afterTextChanged()`.

This slightly modifies backspace behavior, because it does
not automatically delete automatically added characters.

For example, if the content of `ExpiryDateEditText` is "12/2",
previously tapping backspace would update the content to "12".
Now, tapping backspace would result in content being "12/".

Fixes #2035