Fix HTML select option rendering in WebView
Summary
`appcompat:1.1.0` introduced an issue that caused WebViews on API 21
and 22 to crash on creation. There was a known fix for this issue,
but that fix broke rendering of HTML select options. This was because
the fix made the `WebView` use an `Application` context instead of
`Activity` context; `Activity` context is required to correctly render
HTML select options.

Upgrade to `appcompat:1.2.0-rc02` and remove the fix.

See https://issuetracker.google.com/issues/141132133 for more details.

Motivation
Fixes #2656

Testing
Manually verified that WebViews render correctly after upgrading to
`appcompat:1.2.0-rc02` and removing the crash workaround.

Manually verified that HTML select options now render correctly when
clicked.