Fix CardNumberEditText text formatting issue (#2756)
`CardNumberEditText` will only apply formatting to its text if the last
key pressed was not delete. This was tracked
via `StripeEditText#isLastKeyDelete`.

A bug was discovered where pressing delete and then pasting a full
card number would result in the text not being formatted.

Update the logic in `CardNumberEditText`'s `TextWatcher` to handle
pastes by also tracking whether the text change resulted in added
digits (e.g. `""` to `"42"` is considered as added digits;
`"4242"` to `"424"` is not). If digits have been added, ignore the
`isLastKeyDelete` flag.