Remove Google Pay from the custom sheet (#3470)
# Summary
1. In the PaymentOptionsActivity we will make sure not to go to the add fragment if google pay is ready.
2. Moved all google pay button functionality out of the BaseAddFragment and into the PaymentSheetAddFragment

# Motivation
The look of the google pay button makes it feel like the authorization is going to happen and the payment go through at that time, which is not the case in the custom work flow.  By selecting google pay on the list, it is more clear that no payment is going to happen.