Only use StripeBrowserLauncherActivity when Custom Tabs are supported (#3634)
Update logic in `PaymentAuthWebViewContract` to only use
`StripeBrowserLauncherActivity` if Custom Tabs are supported on the
device.

While testing on emulators with older API versions (e.g. API 21)
that didn't have Chrome installed, it was discovered that return URL
handling wasn't working as expected.

This PR reduces the scope of #3596 until we can verify behavior
on non-Chrome devices or add a check for the availability of Chrome.