Simplify construction of view model factories
With Lifecycle 2.5.0, we can simplify the construction of our view models. Instead of subclassing `AbstractSavedStateViewModelFactory` and having to provide an `Application`, `SavedStateRegistryOwner`, and default arguments, we can now conform to `ViewModelProvider.Factory` and override the new `create(modelClass, extras)`.

The `extras` are automatically populated with the `Application` and provide us with a way to create a `SavedStateHandle`. Thus, we no longer need to provide the `Application` or a `SavedStateRegistryOwner` when instantiating the factories.