Add tests for PaymentMethodUI.
We have existing tests for the plural of this (which proves my code change didn't actually affect the display of this).

But I plan to update those tests in a follow up to reduce the amount of permutations we check. The plan is to only test the layout, not the individual payment methods.