Represent missing pattern as null instead of empty string.
This forces us to consider whether the pattern is missing anywhere that we use it, rather than having to remember to check if it's empty or not.

We might throw an IllegalArgumentException in the constructor now, but this should be fine because our tests will fail if we ever create an empty Metadata instance.