[NME] Accepting and sending new api param for acquiring consent for networking manual entry consent (#8972)
* [skip ci] Start PR

* feat: Add acquire_consent_on_primary_cta_click to NetworkedAccountsList

This commit adds a new field `acquire_consent_on
_primary_cta_click` to the `NetworkedAccountsList` model. This field indicates whether consent should be acquired when the primary CTA is clicked.

The field is used in the `LinkAccountPickerViewModel` to determine whether to acquire consent when the user selects accounts. If the field is true, the
 `selectAccounts` function will call the `postShareNetworkedAccounts` repository function with the `consentAcquired` parameter set to true.

* feat: Add acquire_consent_on_primary_cta_click to NetworkedAccountsList

This commit adds a new field `acquire_consent_on
_primary_cta_click` to the `NetworkedAccountsList` model. This field indicates whether consent should be acquired when the primary CTA is clicked.

The field is used in the `LinkAccountPickerViewModel` to determine whether to acquire consent when the user selects accounts. If the field is true, the
 `selectAccounts` function will call the `postShareNetworkedAccounts` repository function with the `consentAcquired` parameter set to true.

* Fixes api param.

* Fixes tests.