[FC] Sends supports_app_verification to sync call when Integrity available (#9819)
# Summary
When making the first `/sync` request, warm up the integrity request calling  manager `prepare` to check if Integrity API is available. If no errors occur, Pass `supports_app_verification: true` on the first sync call.

Note: This gets recorded on the session so even if future sync calls pass this field it'd get ignored, as the session verified state is immutable on backend. 

We also pass `verified_app_id`, matching the package name.

# Motivation
https://docs.google.com/document/d/1joKz5UZHLVazmecfMHbq6gB6n4wj5u8To6AtqYgq_tc/edit?tab=t.0#heading=h.cz1xkpga7giy

# Testing
- [x] Modified tests
- [x] Manually verified